requirements.txt
Django>=4.2.0
celery==5.3.6
redis>=6.2.0
Pillow>=10.0.0
python-dotenv>=1.0.0
opencv-python>=4.8.0
numpy>=1.24.0
scikit-learn>=1.3.0
tensorflow>=2.13.0
django-redis>=5.4.0
psutil>=5.9.6
django-debug-toolbar>=4.2.0
django-extensions>=3.2.3
whitenoise>=6.6.0
gunicorn>=21.2.0
openai>=1.3.7
djangorestframework>=3.14.0
django-filter>=23.0
django-cors-headers>=4.0.0 